Try to use backpack.
It's a pitty that we require a user to user Prometheus
even it's not needed for him. So we want to find a way
to avoid that. It seems that backback may solve this
problem. So the library can be instantiated with a correct
deps that require only the things that are needed.